ANVL: Try to use the distro kernel if dtb available
If there's /boot/anvl.dtb available in eMMC ext4 partition,
let's try to use that and /boot/zImage. Note that currently
the default name for the dtb would be /boot/omap3-anvl.dtb,
but let's not use that as it may appear after updating the
system with incomplete kernel support. To avoid that, let's
require a different name for the dtb.