Fix booting from NAND
Looks like some things need to be updated to boot from NAND
since we've moved the addresses around a bit since the last
time NAND was being used.

Note that we're using ti,nand-ecc-opt = "sw" because that was
the default on earlier kernels. With recent kernels we could
use hardware ecc with ti,nand-ecc-opt = "bch8" but that
requires redoing the rootfs.