Fixes for kernels where eMMC shows up properly at mmc1
Now with eMMC showing up as mmc1 like it should, we need to
use mmcblk1 in bootargs. And since we no longer have NAND,
let's allow saving environment to mmc.

To change the MMC to use as root, just do:

# setenv mmcroot /dev/mmcblk1p2 rw
# saveenv

Where /dev/mmcblk1p2 is for eMMC and /dev/mmcblk0p2 is for
mmc slot.